# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: UIT Career Hub API
  version: 1.0.0
  description: Contract skeleton cho happy flow MVP của một trường UIT.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /students/me:
    get:
      tags: [Students]
      operationId: getMyStudentProfile
      responses:
        '200':
          description: Hồ sơ sinh viên
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentProfileEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      tags: [Students]
      operationId: updateMyStudentProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentProfileUpdateRequest'
      responses:
        '200':
          description: Hồ sơ đã cập nhật
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentProfileEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /students/me/documents/uploads:
    post:
      tags: [Students]
      operationId: createMyDocumentUploadIntent
      description: Tạo URL PUT ký trước để trình duyệt tải PDF trực tiếp vào Cloudflare R2 private.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StudentDocumentUploadRequest' }
      responses:
        '201':
          description: Upload intent có thời hạn
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentDocumentUploadIntentEnvelope' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /students/me/documents/uploads/{uploadId}/complete:
    post:
      tags: [Students]
      operationId: completeMyDocumentUpload
      description: Kiểm tra object trên R2 bằng HEAD, tạo tài liệu ở trạng thái PENDING và thông báo cho UIT xử lý.
      parameters:
        - name: uploadId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201':
          description: Danh sách tài liệu sau khi xác nhận upload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentDocumentListEnvelope' }
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          description: Upload intent đã hết hạn
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /students/me/documents/{documentId}:
    delete:
      tags: [Students]
      operationId: deleteMyDocument
      description: Xóa vĩnh viễn tài liệu khỏi hồ sơ và R2. Không cho xóa CV mặc định hoặc tài liệu đã dùng trong đơn ứng tuyển.
      parameters:
        - name: documentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Danh sách tài liệu còn lại sau khi xóa
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentDocumentListEnvelope' }
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /students/me/documents/{documentId}/download:
    post:
      tags: [Students]
      operationId: createMyDocumentDownloadUrl
      description: Tạo URL GET ký trước ngắn hạn cho tài liệu thuộc sinh viên đang đăng nhập.
      parameters:
        - name: documentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: URL tải xuống có thời hạn
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentDocumentDownloadEnvelope' }
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /students/me/documents/{documentId}/default:
    post:
      tags: [Students]
      operationId: setMyDefaultCv
      description: Chỉ CV đã được UIT xác minh mới có thể được chọn làm CV mặc định.
      parameters:
        - name: documentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Danh sách tài liệu sau khi cập nhật CV mặc định
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StudentDocument'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /uit/student-documents:
    get:
      tags: [UIT]
      operationId: listStudentDocumentsForReview
      description: Danh sách tài liệu sinh viên theo trạng thái, sắp xếp tài liệu cũ nhất trước.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema: { type: string, enum: [PENDING, VERIFIED, REJECTED], default: PENDING }
        - name: query
          in: query
          schema: { type: string, maxLength: 120 }
      responses:
        '200':
          description: Hàng đợi xác minh tài liệu
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UitStudentDocumentListEnvelope' }
        '403':
          $ref: '#/components/responses/Forbidden'
  /uit/student-documents/{documentId}/download:
    post:
      tags: [UIT]
      operationId: createUitStudentDocumentDownloadUrl
      description: Tạo URL GET ký trước ngắn hạn để UIT đối chiếu tài liệu PDF private.
      parameters:
        - name: documentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: URL tải xuống có thời hạn
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentDocumentDownloadEnvelope' }
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /uit/student-documents/{documentId}/review:
    post:
      tags: [UIT]
      operationId: reviewStudentDocument
      description: Chỉ cho phép chuyển một lần từ PENDING sang VERIFIED hoặc REJECTED; quyết định và lý do được lưu audit log.
      parameters:
        - name: documentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StudentDocumentReviewRequest' }
      responses:
        '200':
          description: Tài liệu sau khi xử lý
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/UitStudentDocument' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /students/me/interviews:
    get:
      tags: [Students, Interviews]
      operationId: listMyInterviews
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: scope
          in: query
          schema: { type: string, enum: [upcoming, history, all], default: all }
      responses:
        '200':
          description: Lịch phỏng vấn thuộc sinh viên đang đăng nhập
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewListEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
  /students/me/interviews/{interviewId}/confirm:
    post:
      tags: [Students, Interviews]
      operationId: confirmMyInterview
      description: Xác nhận tự nhiên idempotent; gọi lại khi đã CONFIRMED trả cùng trạng thái và không tạo thông báo trùng.
      parameters:
        - $ref: '#/components/parameters/InterviewId'
      responses:
        '200':
          description: Sinh viên đã xác nhận tham gia và doanh nghiệp được thông báo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewListItemEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /applications:
    get:
      tags: [Applications]
      operationId: listMyApplications
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ApplicationStatus'
      responses:
        '200':
          description: Đơn của sinh viên hiện tại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationListEnvelope'
    post:
      tags: [Applications]
      operationId: submitApplication
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationSubmitRequest'
      responses:
        '201':
          description: Đơn được tạo ở UIT_REVIEWING
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /applications/{applicationId}:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    get:
      tags: [Applications]
      operationId: getMyApplication
      responses:
        '200':
          description: Chi tiết, timeline và hành động hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
  /applications/{applicationId}/offer-document/download:
    post:
      tags: [Applications]
      operationId: createStudentOfferDocumentDownload
      summary: Tạo URL tạm thời để sinh viên mở PDF offer của chính mình
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: URL ký trước có hiệu lực ngắn hạn
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentDocumentDownloadEnvelope' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /applications/{applicationId}/resubmit:
    post:
      tags: [Applications]
      operationId: resubmitApplication
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationResubmitRequest'
      responses:
        '200':
          description: Kiểm tra đúng loại tài liệu UIT yêu cầu, tạo snapshot mới và chuyển NEEDS_SUPPLEMENT sang UIT_REVIEWING
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /applications/{applicationId}/withdraw:
    post:
      tags: [Applications]
      operationId: withdrawApplication
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: Rút đơn tại UIT_REVIEWING, NEEDS_SUPPLEMENT, FORWARDED_TO_COMPANY hoặc COMPANY_REVIEWING
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /applications/{applicationId}/cancel-interview:
    post:
      tags: [Applications, Interviews]
      operationId: cancelInterviewParticipation
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: Hủy lịch đang hoạt động từ INTERVIEW_INVITED, chuyển đơn WITHDRAWN và ghi history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /applications/{applicationId}/offer/accept:
    post:
      tags: [Applications]
      operationId: acceptOffer
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: OFFER_PENDING_STUDENT sang ACCEPTED_PENDING_UIT_CONFIRMATION
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationEnvelope'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cronBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: CRON_SECRET
    refreshCookie:
      type: apiKey
      in: cookie
      name: uit_refresh_token
...
  responses:
    BadRequest:
      description: Request không hợp lệ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Forbidden:
      description: Không có quyền thực hiện
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    ServiceUnavailable:
      description: Tích hợp hạ tầng cần thiết chưa được cấu hình hoặc tạm thời không khả dụng
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    StudentProfileUpdateRequest:
      type: object
      required: [phone]
      additionalProperties: false
      properties:
        phone: { type: [string, 'null'], minLength: 8, maxLength: 30 }
    StudentProfileEnvelope:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/StudentProfile' }
...
    StudentDocumentListEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/StudentDocument' }
...
    StudentDocumentUploadRequest:
      type: object
      additionalProperties: false
      required: [documentType, fileName, mimeType, fileSizeBytes]
      properties:
        documentType: { type: string, enum: [CV, TRANSCRIPT, STUDENT_CONFIRMATION, OTHER] }
        fileName: { type: string, minLength: 1, maxLength: 255, pattern: '\\.[pP][dD][fF]$' }
        mimeType: { type: string, const: application/pdf }
        fileSizeBytes: { type: integer, minimum: 1, maximum: 10485760 }
...